+++ library/Morloc/CodeGenerator/Docstrings.hs
@@ -215,6 +215,7 @@ processArgDoc i (FunT ts t) (ArgDocSig cmddoc argdocs retdoc) = do
       , cmdDocRet = (t', getReturnDesc retdoc' (docReturn cmddoc))
       , cmdDocRetMime = getReturnMime retdoc'
       , cmdDocTerminals = docWith cmddoc
+      , cmdDocEpilogues = docEpilogues cmddoc
       }
 processArgDoc i t (ArgDocSig cmddoc [] retdoc) = do
   loc <- argLocPrefix i
@@ -229,6 +230,7 @@ processArgDoc i t (ArgDocSig cmddoc [] retdoc) = do
       , cmdDocRet = (t', getReturnDesc retdoc' (docReturn cmddoc))
       , cmdDocRetMime = getReturnMime retdoc'
       , cmdDocTerminals = docWith cmddoc
+      , cmdDocEpilogues = docEpilogues cmddoc
       }
 processArgDoc i t (ArgDocAlias r) = do
   loc <- argLocPrefix i
@@ -247,6 +249,7 @@ processArgDoc i t (ArgDocAlias r) = do
       , cmdDocRet = (t', [])
       , cmdDocRetMime = getReturnMime r'
       , cmdDocTerminals = []
+      , cmdDocEpilogues = docEpilogues r
       }
 processArgDoc i t r = do
   (t', r') <- reduceArgDoc i t r
@@ -267,6 +270,7 @@ processArgDoc i t r = do
           , cmdDocRet = (t', [])
           , cmdDocRetMime = Nothing
           , cmdDocTerminals = []
+          , cmdDocEpilogues = []
           }
     _ -> MM.throwSystemError "Expected a record type with docstrings but found a non-record type"
 
@@ -422,6 +426,7 @@ reduceArgDocFrom seen i t@(VarT v) arg
         , docListChecks = if null (docListChecks r1) then docListChecks r2 else docListChecks r1
         , docWith = if null (docWith r1) then docWith r2 else docWith r1
         , docMime = docMime r1 <|> docMime r2
+        , docEpilogues = if null (docEpilogues r1) then docEpilogues r2 else docEpilogues r1
         }
 reduceArgDocFrom seen i (NamT o v ps (map snd -> ts)) (ArgDocRec arg rs) = do
   let args = map (ArgDocAlias . snd) rs